requirements.txt
flask
pandas
pymongo
python-dotenv
openai
numpy